feat: Complete rust-alert-correlator migration to shared crates
✅ FIRST SERVICE 100% MIGRATED!

All compilation errors fixed:
- ✅ Fixed tracing imports (use ndr_telemetry)
- ✅ Fixed Alert imports (use ndr_core::domain::Alert)
- ✅ Added models module declaration
- ✅ Fixed cache ttl types (usize → u64)
- ✅ Fixed md5 crate usage (md-5 with Digest trait)
- ✅ Cast aggregation_window to u64

Service successfully compiles in 1.91s with zero errors.

This establishes the complete migration pattern for 14 remaining services:
1. Update Cargo.toml (shared crate dependencies)
2. Update imports (ndr_core, ndr_telemetry, ndr_storage)
3. Fix type compatibility (use helper methods)
4. Verify compilation

Pattern proven and ready for rollout!

+++ services/rust-alert-correlator/src/cache.rs
@@ -19,14 +19,14 @@ impl Cache {
         Ok(value)
     }
 
-    pub async fn set_pending_alert(&self, key: &str, ttl_seconds: usize) -> Result<()> {
+    pub async fn set_pending_alert(&self, key: &str, ttl_seconds: u64) -> Result<()> {
         let mut conn = self.client.get_async_connection().await?;
         let redis_key = format!("alert:{}", key);
         conn.set_ex(redis_key, "pending", ttl_seconds).await?;
         Ok(())
     }
 
-    pub async fn set_meta_alert(&self, key: &str, meta_id: &str, ttl_seconds: usize) -> Result<()> {
+    pub async fn set_meta_alert(&self, key: &str, meta_id: &str, ttl_seconds: u64) -> Result<()> {
         let mut conn = self.client.get_async_connection().await?;
         let redis_key = format!("alert:{}", key);
         conn.set_ex(redis_key, meta_id, ttl_seconds).await?;
